+++ js/multiselect.js
@@ -22,7 +22,8 @@
         defaults = {
           filterMode: 'contains',
           maxSelected: undefined,
-          moveSelectedToTop: true,
+          moveSelected: 'all',
+          showEmptyGroupHeaders: false,
           source: undefined
         },
         settings = $.extend({}, defaults, options);
@@ -50,6 +51,7 @@
         var ddOpts = {
             closeOnSelect: false,
             empty: true,
+            moveSelected: 'all',
             multiple: true
           };
 
@@ -65,8 +67,12 @@
           ddOpts.maxSelected = this.settings.maxSelected;
         }
 
-        if (this.settings.moveSelectedToTop) {
-          ddOpts.moveSelectedToTop = this.settings.moveSelectedToTop;
+        if (this.settings.moveSelected) {
+          ddOpts.moveSelected = this.settings.moveSelected;
+        }
+
+        if (this.settings.showEmptyGroupHeaders) {
+          ddOpts.showEmptyGroupHeaders = this.settings.showEmptyGroupHeaders;
         }
 
         this.element.dropdown(ddOpts);